# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Workforce API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/workers": {
      "get": {
        "summary": "Get all workers",
        "responses": {
          "200": {
            "description": "A list of workers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/worker"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new worker",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/worker"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The created worker",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/worker"
                }
              }
            }
          }
        }
      }
    },
    "/workers/{id}": {
      "get": {
        "summary": "Get a worker by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The worker",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/worker"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a worker by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/worker"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated worker",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/worker"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a worker by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The deleted worker",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/worker"
                }
              }
            }
          }
        }
      }
    },
...
    "/skills": {
      "get": {
        "summary": "Get all skills",
        "responses": {
          "200": {
            "description": "A list of skills",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/skill"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new skill",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/skill"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The created skill",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/skill"
                }
              }
            }
          }
        }
      }
    },
    "/skills/{id}": {
      "get": {
        "summary": "Get a skill by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The skill",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/skill"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a skill by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/skill"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated skill",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/skill"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a skill by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The deleted skill",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/skill"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "skill": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "description"
        ]
      },
...
      "worker": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "The type of the worker",
            "enum": [
              "ai-worker",
              "human-worker"
            ]
          },
          "credential": {
            "type": "string"
          },
          "channelUserConfig": {
            "type": "object",
            "description": "A map of channel types to channel user credential names",
            "properties": {
              "slack-channel": {
                "type": "string"
              },
              "native-channel": {
                "type": "string"
              },
              "discord-channel": {
                "type": "string"
              }
            }
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "The list of skills the worker has."
          },
          "wipLimit": {
            "type": "number",
            "description": "The number of tasks a worker can have in progress at a time."
          },
          "variables": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/ai_worker_variables"
              },
              {
                "$ref": "#/components/schemas/human_worker_variables"
              }
            ]
          }
        },
        "required": [
          "type",
          "credential",
          "variables"
        ]
      },
...
      "ai_worker_variables": {
        "type": "object",
        "properties": {
          "model": {
            "type": "string",
            "description": "The name of the model to use for inference",
            "default": "gpt-3.5-turbo",
            "enum": [
              "gpt-3.5-turbo",
              "gpt-3.5-turbo-instruct",
              "gpt-3.5-turbo-16k",
              "gpt-4",
              "gpt-4-32k",
              "gpt-4-0125-preview",
              "gpt-4o",
              "claude-3-5-sonnet-20240620",
              "claude-3-opus-20240229",
              "claude-3-haiku-20240307",
              "llama3-8b-8192",
              "llama3-70b-8192",
              "llama2-70b-4096",
              "mixtral-8x7b-32768",
              "gemma-7b-it"
            ]
          },
          "temperature": {
            "type": "number",
            "description": "The temperature parameter of the model, lower values are more deterministic, higher are more random.",
            "default": 0,
            "minimum": 0,
            "maximum": 1
          },
          "max_tokens": {
            "type": "number",
            "description": "The maximum number of tokens to generate",
            "default": 2048,
            "minimum": 1,
            "maximum": 128000
          },
          "top_p": {
            "type": "number",
            "description": "The top P value to use.  Higher values are more deterministic, lower values are more chaotic.",
            "default": 1,
            "minimum": 0,
            "maximum": 1
          }
        },
        "required": [
          "model"
        ]
      },
      "human_worker_variables": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "string",
            "description": "The user id of the worker"
          }
        },
        "required": [
          "user_id"
        ]
      },
...
    }
  }
}